requirements.txt
gradio
numpy
python-dotenv
sentence-transformers
scikit-learn
pdfplumber
python-docx
openai
ollama